fix all error compil

+++ includes/game.h
@@ -158,7 +158,7 @@ typedef struct s_player
 	float		f_tel;
 	bool		invert_mouse_x;
 	int			selected_anim;
-	int			animation;
+	float		animation;
 	t_raycast	*raycast;
 	t_stats		*stats;
 	t_keycode	*key;
